+++ src/main/java/tave/crezipsa/crezipsa/application/community/usecase/LikeUseCaseImpl.java
@@ -1,7 +1,5 @@
 package tave.crezipsa.crezipsa.application.community.usecase;
 
-import java.time.LocalDateTime;
-
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
@@ -99,19 +97,4 @@ public class LikeUseCaseImpl  implements LikeUseCase {
 		});
 	}
 
-	public static String preview(String content) {
-		return content.substring(0, Math.min(50, content.length()));
-	}
-
-	public static String convertToRelativeTime(LocalDateTime createdAt) {
-		LocalDateTime now = LocalDateTime.now();
-		long minutes = java.time.Duration.between(createdAt, now).toMinutes();
-		long hours = minutes / 60;
-		long days = hours / 24;
-
-		if (minutes < 1) return "방금 전";
-		if (minutes < 60) return minutes + "분 전";
-		if (hours < 24) return hours + "시간 전";
-		return days + "일 전";
-	}
 }
